Refactor local copy options into dedicated module
